# src/main/resources/openapi_napolya.yaml
openapi: 3.0.3
info:
  title: API na_polya 3.0
  version: 1.0.11
  description: Все необходимые API для сайта На_поля
...
paths:
...
  /notifications:
    get:
      tags:
        - Notification
      summary: Получить список id уведомлений
      responses:
        '200':
          description: Успешный ответ с массивом id уведомлений
          content:
            application/json:
              schema:
                type: array
                items:
                  type: integer
                  description: Список id уведомлений
                  example: [1, 2]
    post:
      tags:
        - Notification
      summary: Создать новое уведомление
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Notification'
      responses:
        '201':
          description: Уведомление успешно создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
  /notifications/{id}:
    get:
      tags:
        - Notification
      summary: Получить информацию об уведомлении
      parameters:
        - name: id
          in: path
          required: true
          description: ID уведомления
          schema:
            type: integer
      responses:
        '200':
          description: Информация об уведомлении
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        '404':
          description: Уведомление не найдено

    put:
      tags:
        - Notification
      summary: Обновить информацию об уведомлении
      parameters:
        - name: id
          in: path
          required: true
          description: ID уведомления
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Notification'
      responses:
        '200':
          description: Уведомление успешно обновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        '404':
          description: Уведомление не найдено
    
    delete:
      tags:
        - Notification
      summary: Удалить уведомление
      parameters:
        - name: id
          in: path
          required: true
          description: ID уведомления
          schema:
            type: integer
      responses:
        '204':
          description: Уведомление успешно удалено
        '404':
          description: Уведомление не найдено
components:
  schemas:
...
    Notification: # Определение сущности "Уведомление"
      type: object
      properties:
        id:
          type: integer
          description: Уникальный идентификатор уведомления
          example: 1
        text:
          type: string
          description: Текст уведомления
          example: "Вы записаны в команду A!"
        type:
          type: string
          description: Тип уведомления
          # Написать список возможных типов
          example: "уведомление о составе команды"
        recipient_id:
          type: integer
          description: id получателя уведомления
          example: 2
        status:
          type: string
          description: Статус уведомления
          enum: [непрочитанное, прочитанное]
          example: "непрочитанное"
        date:
          type: string
          description: Дата и время отправки уведомления
          format: date-time
          example: "2024-11-16T10:00:00Z"

      required:
        - id
        - text
        - type
        - recipient_id
        - status
        - date